correção nota avaliacao pra double

+++ src/main/java/br/com/villaca/arte/dto/request/AvaliacaoRequest.java
@@ -2,8 +2,8 @@ package br.com.villaca.arte.dto.request;
 
 import java.time.Instant;
 
-public record AvaliacaoRequest(        
-        Integer nota,
+public record AvaliacaoRequest(
+        Double nota,
         String comentario,
         Instant dataAvaliacao,
         UsuarioRequest usuario,
